requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
waitress==2.1.2
numpy-stl
requests
numpy
pandas
scikit-learn
scipy
scikit-image

